requirements.txt
Flask
requests
python-dotenv
slack_bolt
slack_sdk
gunicorn